Added full support for translations phases 1 & 2 using new mstr object.
The mstr object allows the code to temporarily parse trigraphs and
backslash-newline pairs. If the line is then edited, these items are
restored when the line is written to output. It's a fair bit of extra
work for a feature that probably no one will ever want or need, sadly.